Cannot delete account with vaults in them

+++ account/src/main/java/com/pgr3402/csgobank/account/AccountService.java
@@ -1,33 +1,44 @@
 package com.pgr3402.csgobank.account;
 
 
+import com.fasterxml.jackson.databind.JsonNode;
 import com.pgr3402.csgobank.account.event.AccountEventEnum;
 import com.pgr3402.csgobank.account.event.AccountEventPub;
 import com.pgr3402.csgobank.transaction.Transaction;
 import com.pgr3402.csgobank.transaction.TransactionState;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.lang.reflect.Field;
-import java.nio.channels.FileChannel;
+import java.util.List;
 import java.util.Optional;
 
 @Service
+@Slf4j
 public class AccountService {
 
     private final AccountRepository accountRepository;
     private final AccountEventPub accountEventPub;
 
-    public AccountService(AccountRepository accountRepository, AccountEventPub accountEventPub) {
+    private final VaultClient vaultClient;
+
+    public AccountService(AccountRepository accountRepository, AccountEventPub accountEventPub, VaultClient vaultClient) {
         this.accountRepository = accountRepository;
         this.accountEventPub = accountEventPub;
+        this.vaultClient = vaultClient;
     }
 
 
-    public void deleteAccount(Account account){
+    public void deleteAccount(Account account) {
         long accountId = account.getId();
-         accountRepository.delete(account);
-         accountEventPub.publishAccountEvent(accountId,AccountEventEnum.DELETED);
+        ResponseEntity<List<JsonNode>> response = vaultClient.getVault(accountId);
+        List<JsonNode> vault = response.getBody();
+        if(vault == null || vault.size() > 0){
+            return;
+        }
+        accountRepository.delete(account);
+        accountEventPub.publishAccountEvent(accountId, AccountEventEnum.DELETED);
     }
 
     public Account createAccount(Account account) {
@@ -71,11 +82,11 @@ public class AccountService {
         return Optional.of(accountRepository.save(updatedAccount));
     }
 
-    public Transaction transferCredits(Transaction transaction){
+    public Transaction transferCredits(Transaction transaction) {
         Optional<Account> optionalFromAccount = accountRepository.findById(transaction.getFromAccountId());
         Optional<Account> optionalToAccount = accountRepository.findById(transaction.getToAccountId());
 
-        if(optionalFromAccount.isEmpty() || optionalToAccount.isEmpty()){
+        if (optionalFromAccount.isEmpty() || optionalToAccount.isEmpty()) {
             transaction.setState(TransactionState.FAILED);
 
             return transaction;
@@ -102,7 +113,7 @@ public class AccountService {
     public Optional<Account> depositCredit(long accountId, int amount) {
         Optional<Account> optionalAccount = accountRepository.findById(accountId);
 
-        if (optionalAccount.isEmpty()){
+        if (optionalAccount.isEmpty()) {
             return Optional.empty();
         }
         Account account = optionalAccount.get();
@@ -115,12 +126,12 @@ public class AccountService {
 
         Optional<Account> optionalAccount = accountRepository.findById(accountId);
 
-        if (optionalAccount.isEmpty()){
+        if (optionalAccount.isEmpty()) {
             return Optional.empty();
         }
         Account account = optionalAccount.get();
 
-        if(account.getCredit() < amount){
+        if (account.getCredit() < amount) {
             return Optional.empty();
         }
         account.withdraw(amount);
